Replace int64 prints with workaround based on int32 prints. Some ESP platforms doesn't support printing 64 bit integers and they crash on %lld, %llu, PRI.64 etc. RemoteDeviceUpdate: fix missing handling of firmware update structure. Logging: for esp32 skip Supla logs and directly call ESP_LOGx methods. check_cpplint.sh: add check for 64 bit prints, so it will fail if somebody will try to add it

+++ src/supla/sensor/general_purpose_channel_base.cpp
@@ -188,7 +188,9 @@ void GeneralPurposeChannelBase::setDefaultValueMultiplier(int32_t multiplier) {
 }
 
 void GeneralPurposeChannelBase::setDefaultValueAdded(int64_t added) {
-  SUPLA_LOG_DEBUG("GPM[%d]: DefaultValueAdded %lld", getChannelNumber(), added);
+  SUPLA_LOG_DEBUG("GPM[%d]: DefaultValueAdded %d",
+                  getChannelNumber(),
+                  static_cast<int>(added));
   defaultValueAdded = added;
 }
